+++ R/fetch_uniprot.R
@@ -77,7 +77,7 @@ fetch_uniprot <-
       dplyr::mutate(accession = stringr::str_extract_all(.data$input_id,
         pattern = "[OPQ][0-9][A-Z0-9]{3}[0-9]|[A-NR-Z][0-9]([A-Z][A-Z0-9]{2}[0-9]){1,2}"
       )) %>%
-      tidyr::unnest(.data$accession)
+      tidyr::unnest("accession")
     uniprot_ids_filtered <- c(uniprot_ids_filtered, valid_id_annotations$accession)
     non_identifiable_id <- non_conform_ids[!stringr::str_detect(non_conform_ids,
       pattern = "[OPQ][0-9][A-Z0-9]{3}[0-9]|[A-NR-Z][0-9]([A-Z][A-Z0-9]{2}[0-9]){1,2}"
@@ -164,7 +164,7 @@ They were fetched and the original input ID can be found in the "input_id" colum
         NA
       )) %>%
       dplyr::distinct(.data$accession, .data$new) %>%
-      tidyr::drop_na(.data$new)
+      tidyr::drop_na("new")
 
     new_ids <- new$new