+++ NEWS.md
@@ -1,4 +1,4 @@
-# protti 0.5.0.9000
+# protti 0.6.0
 
 ## New features